django/requirements.txt
beautifulsoup4
Django
django-tz-detect
python-dateutil
requests
django-import-export
Pillow
psycopg2
gunicorn